Task 5 - done
Implement the Book class.
It is necessary to store in the class fields: title of the book, full name of the author, year of publication, name of the publisher, genre of the book, number of pages.
Implement class constructors and methods for data input, data output, and implement access to individual fields through class methods.
Use method overloading mechanism.